LayoutTests/imported/w3c/web-platform-tests/workers/same-site-cookies/third-party.default.tentative.sub.https.window-expected.txt


FAIL Check SharedWorker sameSiteCookies option default for third-party assert_equals: Worker should get/set SameSite=None cookies only expected "ReadOnLoad:None,ReadOnFetch:None,SetOnRedirectLoad:None,SetOnLoad:None,SetOnRedirectFetch:None,SetOnFetch:None" but got "ReadOnLoad:StrictLaxNone,ReadOnFetch:StrictLaxNone,SetOnRedirectLoad:StrictLaxNone,SetOnLoad:StrictLaxNone,SetOnRedirectFetch:StrictLaxNone,SetOnFetch:StrictLaxNone"
